Refactor CommunityService imports to pageservices package for improved organization
