+++ src/controllers/Controller.java
@@ -1,7 +1,5 @@
 package src.controllers;
 
-import java.sql.SQLException;
-
 public interface Controller {
     void setOnEvent();
 }
