+++ test/unittest/UnitTest.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import src.services.*;
 import src.services.componentservices.CommentService;
 import src.services.componentservices.PostService;
 import src.services.pageservices.CommunityCreationPageService;
+import src.services.pageservices.CommunityService;
 import src.services.pageservices.PostCreationPageService;
 import src.domainmodel.User;
 import src.persistence.dbmanager.DBConnection;
